refactor : 거리요금타입 fare 패키지로 이동
